backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-cloud-modelarmor
google-cloud-secret-manager
psycopg2-binary
requests
python-dotenv
pydantic
sse-starlette
